DPB 2.2 Repaired by Tony "omega" Sergi.
There are solution files for VC2003.Net now.
You may need to check with the configs inside vc to repair the linux makefile, as i didn't touch it. but the server _should_ still compile in linux, as is.
i didn't add any new .cpp files on the server as far as i can remember, so it should still compile as it did, without needing to modify the makefile. i hope! haha